+++ ee-express-core/src/main/java/de/simonkerstan/ee/core/configuration/SystemPropertiesConfigurationProvider.java
@@ -5,16 +5,22 @@
 
 package de.simonkerstan.ee.core.configuration;
 
+import java.util.List;
 import java.util.Optional;
 
 /**
  * Configuration properties from the system properties (automatically loaded by the JVM).
  */
-public class SystemPropertiesConfigurationProvider implements ConfigurationProvider {
+public final class SystemPropertiesConfigurationProvider implements ConfigurationProvider {
 
     @Override
     public Optional<String> getConfigurationValue(String propertyName) {
         return Optional.ofNullable(System.getProperty(propertyName));
+    }
+
+    @Override
+    public Optional<List<String>> getConfigurationSubValues(String propertyName) {
+        return Optional.empty();
     }
 
 }