+++ CHANGELOG.md
@@ -12,10 +12,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - EE Express Persistence (Jakarta EE Persistence, Jakarta EE Data and Redis support)
 - Configuration parsing for record types
 - Missing primitive types to the configuration parsing mechanism
+- Missing configuration file sources
+- Support for configuration lists and maps
 
 ### Changed
 
 - Made `de.simonkerstan:ee-express` publishable via Maven Central.
+- Added required methods to the `ConfigurationProvider` interface.
 
 ## [0.1.0] - 2025-09-05